fix(k8s): improve error handling for Dask K8s resources (#618)
Closes #617
